dump: Extract tar code to internal/dump
