+++ internal/dump/acl_test.go
@@ -1,4 +1,4 @@
-package main
+package dump
 
 import (
 	"reflect"
